pyproject.toml
[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"

[project]
name = "renderctl"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.12",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0",
]

[project.scripts]
renderctl     = "renderctl.cli:app"
renderctl-mcp = "renderctl.mcp_server:main"

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools.packages.find]
include = ["renderctl*"]
